requirements.txt
numpy==1.26.4
torch==2.7.1
scipy==1.13.1
tqdm==4.67.1
openai==1.93.0
python-dotenv==1.1.1
google-generativeai==0.8.5
scikit-learn==1.7.0
PyYAML==6.0.2
gensim==4.3.3
